requirements.txt
mock
pytest
requests
readability-lxml
beautifulsoup4
sentence-transformers
chromadb
langchain
langchain-ollama
langchain-huggingface
langchain-chroma
yt-dlp 